get_check_data fix declaration

+++ src/common/get_check_data.c
@@ -270,7 +270,7 @@ spi_getBigIntArr(
     HeapTuple *tuple,
     TupleDesc *tupdesc,
     Column_info_t info,
-    uint64_t *the_size) {
+    size_t *the_size) {
   bool is_null = false;
 
   Datum raw_array = SPI_getbinval(*tuple, *tupdesc, info.colNumber, &is_null);
